freertosのテストを追加
https://qiita.com/azuki_bar/items/d84e504fcd32dddc5c4b
を参考にfreertosのテストを追加
