+++ training/1_Task_Static/TaskCreateStatic.h
@@ -1,0 +1,6 @@
+#ifndef TASKCREATESTATIC_H
+#define TASKCREATESTATIC_H
+
+extern void TaskCreateStatic( void );
+
+#endif /* TASKCREATESTATIC_H */
